resource pullzone_hostname: handle internal hostnames transparently
internal hostnames (`*.b-cdn.net`) are automatically created when a pullzone is created and cannot be deleted

+++ internal/provider/resource_pullzone_hostname.go
@@ -265,6 +265,11 @@ func (r *PullzoneHostnameResource) Delete(ctx context.Context, req resource.Dele
 		return
 	}
 
+	// b-cdn.net hostnames cannot be deleted
+	if data.IsInternal.ValueBool() {
+		return
+	}
+
 	err := r.client.DeletePullzoneHostname(data.PullzoneId.ValueInt64(), data.Name.ValueString())
 	if err != nil {
 		resp.Diagnostics.Append(diag.NewErrorDiagnostic("Error deleting hostname", err.Error()))
